Rename package node to session
This is part of renaming the top level project from "dst-go" to
"perun-node" and is done as the name "node' is required for the
top level package.

Relates to #55
