+++ session/config.go
@@ -14,7 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package node
+package session
 
 // WalletConfig defines the parameters required to configure a wallet.
 type WalletConfig struct {
